refactor: :recycle: Seperated DOM manipulation form leaflet.ts into domManipulation.ts
