+++ src/leaflet.ts
@@ -2,10 +2,7 @@ import L, { marker } from "leaflet"
 import "leaflet/dist/leaflet.css";
 import "../assets/styles/index.scss";
 import { fromEvent, Subject, takeUntil } from "rxjs";
-import { map } from "rxjs/operators";
 import { Flight } from "./model/Flight";
-
-const cleaup$ = new Subject();
 
 const markers: L.Marker[] = [];
 
@@ -20,19 +17,12 @@ L.tileLayer(
     }
 ).addTo(leafletMap);
 
+// Funcitons to render flight Markers
 export function renderMap(flights: Flight[]){
 
-    // move to dom ts
-    cleaup$.next(true);
-
+    // Clear layer if exists
     if(leafletMap && markers.length > 0){
         markers.forEach(marker => leafletMap.removeLayer(marker))
-    }
-
-    // move to dom ts
-    const flightListElement = document.getElementById(`flights_list`)
-    if (!!flightListElement) {
-        flightListElement.innerHTML = ``;
     }
 
     // Add markers
@@ -61,40 +51,7 @@ export function renderMap(flights: Flight[]){
           );
 
         markers.push(marker);
-
-
-        // move to dom ts
-        // Add flight item to flights list
-        const flightItem = document.createElement(`div`);
-
-        // Item styling
-        flightItem.className = 
-        `flight-item bg-jawgdark-500 p-4 mb-4 mt-6
-        hover:cursor-pointer hover:bg-jawgdark-400 hover:shadow-white-glow
-        lg:z-10 lg:transition-all lg:ease-in-out lg:delay-50 lg:duration-200
-        lg:hover:translate-x-1 lg:hover:w-[405px]
-        border-solid border-1 rounded-md
-        drop-shadow-lg`;
-
-        flightItem.id = flight.icao24address;
-        flightItem.innerHTML = `
-        <div class="flight-item-header hover:shadow-none shadow-white-glow text-black rounded-md bg-jawgdark-50 p-1 text-center -mt-7 w-[50%] drop-shadow-md">
-            <strong>Flight</strong> - ${flight.callsign}
-        </div>
-        <div class="pt-2 text-jawgdark-50">
-            <strong>Origin</strong> - ${flight.origin_country}
-        </div>`;
-        document.getElementById(`flights_list`)!.appendChild(flightItem);
     });
-
-    // Setup eventListeners
-    document.querySelectorAll(`.flight-item`).forEach((item) => {
-        const matchedFlight = flights.find(
-          (flight) => flight.icao24address === item.id
-        );
-        if (matchedFlight) {
-          fromEvent(item, 'click').pipe(takeUntil(cleaup$)).subscribe(() => leafletMap.flyTo([matchedFlight.latitude, matchedFlight.longitude], 8));
-        }
-      });
-
 }
+
+export function getMap(){return leafletMap}
